Added URL for data source
Full dataset not added as this violates the terms of service of Kaggle. To retrieve the dataset, open employee_attrition.url the click on "download". Dataset is a CSV file named "WA_Fn-UseC_-HR-Employee-Attrition". This data is from a 2014 survey that measures attitudes towards mental health and frequency of mental health disorders in the tech workplace using several factors such as income, job role and education